CDAP-15823 package renamed and migrated to CDAP core 6.1.0-SNAPSHOT
