+++ pom.xml
@@ -21,7 +21,7 @@
 
   <groupId>io.cdap.plugins</groupId>
   <artifactId>elasticsearch-plugins</artifactId>
-  <version>1.10.0</version>
+  <version>1.10.0-SNAPSHOT</version>
 
   <licenses>
     <license>
@@ -80,10 +80,10 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <cdap.version>6.0.0-SNAPSHOT</cdap.version>
+    <cdap.version>6.1.0-SNAPSHOT</cdap.version>
     <hadoop.version>2.3.0</hadoop.version>
     <junit.version>4.11</junit.version>
-    <hydrator.version>2.2.0-SNAPSHOT</hydrator.version>
+    <hydrator.version>2.3.0-SNAPSHOT</hydrator.version>
     <es.version>5.4.0</es.version>
     <es-hadoop.version>5.4.0</es-hadoop.version>
     <slf4j.version>1.7.5</slf4j.version>
@@ -96,8 +96,6 @@
     <!-- this is here because project.basedir evaluates to null in the script build step -->
     <main.basedir>${project.basedir}</main.basedir>
   </properties>
-
-
 
   <dependencies>
     <dependency>
@@ -326,9 +324,10 @@
           <extensions>true</extensions>
           <configuration>
             <instructions>
-              <_exportcontents>io.cdap.plugin.batch.*;org.elasticsearch.hadoop.mr.*;
-                org.apache.commons.lang;org.apache.commons.*;org.apache.commons.logging.*;
-                org.codehaus.jackson.*;io.netty.*</_exportcontents>
+              <_exportcontents>
+                io.cdap.plugin.elastic.*;
+                org.elasticsearch.hadoop.mr.*;
+              </_exportcontents>
               <Embed-Dependency>*;inline=false;scope=compile</Embed-Dependency>
               <Embed-Transitive>true</Embed-Transitive>
               <Embed-Directory>lib</Embed-Directory>
@@ -349,8 +348,8 @@
           <version>1.1.0</version>
           <configuration>
             <cdapArtifacts>
-              <parent>system:cdap-data-pipeline[5.0.0,6.0.0-SNAPSHOT)</parent>
-              <parent>system:cdap-data-streams[5.0.0,6.0.0-SNAPSHOT)</parent>
+              <parent>system:cdap-data-pipeline[6.1.0-SNAPSHOT,7.0.0-SNAPSHOT)</parent>
+              <parent>system:cdap-data-streams[6.1.0-SNAPSHOT,7.0.0-SNAPSHOT)</parent>
             </cdapArtifacts>
           </configuration>
           <executions>
